AMC: some changes to the code according to the Aug paper revision
NOTE: this is still WiP - specifically, the implementation of the
initial exploration phase is still buggy.

Accuracy-guaranteed compression finds solutions that have ~25% less
compute (VGG16-CIFAR), but only if the first layer is skipped.  I
expect the Agent to learn that layer #1 is extra-sensitive.  No
use in debugging this until fixing the integration with Coach.