Integrate Cadene pretrained PyTorch models [fixes #142] (#184)
Integrate Cadene ```pretrainedmodels``` package.

This PR integrates a large set of pre-trained PyTorch image-classification and object-detection models which originate from https://github.com/Cadene/pretrained-models.pytorch.

*******************************************************************************************
PLEASE NOTE: 
This PR adds a dependency on he ```pretrainedmodels``` package, and you will need to install it using ```pip3 install pretrainedmodels```.  For new users, we have also updated the ```requirements.txt``` file.
*******************************************************************************************

Distiller does not currently support the compression of object-detectors (a sample application is required - and the community is invited to send us a PR).

Compression of some of these models may not be fully supported by Distiller due to bugs and/or missing features.  If you encounter any issues, please report to us.

Whenever there is contention on the names of models passed to the ```compress_classifier.py``` sample application, it will prefer to use the Cadene models at the lowest priority (e.g. Torchvision models are used in favor of Cadene models, when the same model is supported by both packages).

This PR also:
* Adds documentation to ```create_model```
* Adds tests for ```create_model```