Greedy filter pruning: add mobilenet_v1 greedy pruning
This is discussed in issue #282, although there @Bowenwu1 was
interested in mobilenet for CIFAR, not ImageNet.

Note that the implementation of the Greedy filter pruning algorithm is
not generic (but it is easily extensible) and supports only a subset
of the models.

An example invocation:
time python3 compress_classifier.py --arch=mobilenet PATH-TO-IMAGENET_DS  --resume=mobilenet_sgd_68.848.pth.tar --greedy --greedy-target-density=0.5 --vs=0 -p=50 --lr=0.1 --gpu=0 --greedy-pruning-step=0.15 --effective-train-size=0.01