Model thinning bug fix
This bug is triggered (for example) when you execute this example code:
python3 compress_classifier.py --arch resnet20_cifar  ../../../data.cifar10 -p=50 --lr=0.3 --epochs=180 --compress=../agp-pruning/resnet20_filters.schedule_agp_4.yaml --resume-from=../ssl/checkpoints/checkpoint_trained_dense.pth.tar --reset-optimizer --vs=0

The root-cause is a (non-functional) code refactoring made in commit
992291cfe4a2d917eef6f926f58dc10a68f82105.

The problem is in an `if` statement handling the Optimizer reshaping.
The bad commit combined two `if` statements into one, and moved this
combined statement to a place where the control flows thru it when
it shouldn't.
The fix reverts back to the two original (and separate) `if` statements.